Refactor flask app into blueprints for segment function
