+++ templates/edit-user-role.html
@@ -13,7 +13,7 @@
           {% endif %}
         {% endwith %}
 
-        <form action="{{url_for('changue_role', user_id = user.id)}}" method="post">
+        <form action="{{url_for('users.changue_role', user_id = user.id)}}" method="post">
           <p class="form-label">Selecciona el rol de {{user.name}}:</p>
           {%for role in roles:%}
           <input type="radio" id="color1" name="user_role" value="{{role.id}}" class="form-check-input">
