+++ templates/header.html
@@ -30,7 +30,7 @@
   <nav class="navbar navbar-expand-lg navbar-light mb-5" id="mainNav" style="background-color: #4C5B61;">
     <div class="container px-4 px-lg-5">
       {% if current_user.is_authenticated %}
-      <a class="navbar-brand" href="{{ url_for('get_all_posts') }}">Inicio</a>
+      <a class="navbar-brand" href="{{ url_for('bills.get_all') }}">Inicio</a>
       {%endif%}
       <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarResponsive"
         aria-controls="navbarResponsive" aria-expanded="false" aria-label="Toggle navigation">
@@ -40,16 +40,13 @@
       <div class="collapse navbar-collapse" id="navbarResponsive">
         <ul class="navbar-nav ms-auto py-4 py-lg-0">
           <!-- Only show Login/Register if user is logged out. Otherwise show "Log Out" -->
-          <!-- <li class="nav-item">
-            <a class="nav-link px-lg-3 py-lg-4" href="{{ url_for('get_all_posts') }}">INICIO</a>
-          </li> -->
           {% if current_user.is_authenticated %}
           <li class="nav-item">
-            <a class="nav-link px-lg-3 py-3 py-lg-4" href="{{ url_for('get_all_posts') }}">Registros</a>
+            <a class="nav-link px-lg-3 py-3 py-lg-4" href="{{ url_for('bills.get_all') }}">Registros</a>
           </li>
           {%if current_user.role.can_view_tags%}
           <li class="nav-item">
-            <a class="nav-link px-lg-3 py-3 py-lg-4" href="{{ url_for('get_post_tag') }}">
+            <a class="nav-link px-lg-3 py-3 py-lg-4" href="{{ url_for('tags.get_post_tag') }}">
               Etiquetas de Documento
             </a>
           </li>
@@ -57,40 +54,40 @@
           
           {%if current_user.role.can_manage_document_types%}
           <li class="nav-item">
-            <a class="nav-link px-lg-3 py-3 py-lg-4" href="{{ url_for('get_post_document_types') }}">
+            <a class="nav-link px-lg-3 py-3 py-lg-4" href="{{ url_for('doc_types.get_post_document_types') }}">
             Tipos de Documento
             </a>
           </li>
           {%endif%}
           {%if current_user.role.can_view_users%}
           <li class="nav-item">
-            <a class="nav-link px-lg-3 py-3 py-lg-4" href="{{ url_for('users_panel') }}">
+            <a class="nav-link px-lg-3 py-3 py-lg-4" href="{{ url_for('users.users_panel') }}">
               Usuarios
             </a>
           </li>
           {%endif%}
           {%if current_user.role.can_view_roles%}
           <li class="nav-item">
-            <a class="nav-link px-lg-3 py-3 py-lg-4" href="{{ url_for('roles_panel') }}">
+            <a class="nav-link px-lg-3 py-3 py-lg-4" href="{{ url_for('roles.roles_panel') }}">
               Roles
             </a>
           </li>
           {%endif%}
 
           <li class="nav-item">
-            <a class="nav-link px-lg-3 py-3 py-lg-4" href="{{ url_for('logout') }}">
+            <a class="nav-link px-lg-3 py-3 py-lg-4" href="{{ url_for('auth.logout') }}">
               Salir
             </a>
           </li>
           {%else:%}
           <!-- Hi {{ current_user.name }}! -->
           <li class="nav-item">
-            <a class="nav-link px-lg-3 py-3 py-lg-4" href="{{ url_for('login') }}">
+            <a class="nav-link px-lg-3 py-3 py-lg-4" href="{{ url_for('auth.login') }}">
               Login
             </a>
           </li>
           <li class="nav-item">
-            <a class="nav-link px-lg-3 py-3 py-lg-4" href="{{ url_for('register') }}">
+            <a class="nav-link px-lg-3 py-3 py-lg-4" href="{{ url_for('auth.register') }}">
               Registrar 
             </a>
           </li>
