+++ templates/register.html
@@ -29,7 +29,7 @@
             {% endfor %}
           {% endif %}
         {% endwith %}
-        <form method="POST" action="{{url_for('register')}}">
+        <form method="POST" action="{{url_for('auth.register')}}">
           {{ form.hidden_tag() }}
           <div class="form-group">
               <strong>
@@ -51,7 +51,7 @@
           </div>
           <div class="d-flex justify-content-start mt-3">
             {{ form.submit(class="btn btn-primary") }}
-            <a href="{{url_for('login')}}" class="btn btn-danger mx-3">Cancelar</a>
+            <a href="{{url_for('auth.login')}}" class="btn btn-danger mx-3">Cancelar</a>
           </div>
                   
         </form>
