+++ templates/roles.html
@@ -28,11 +28,11 @@
           </a>
           <p class="post-meta">            
             {%if current_user.role.can_edit_roles%}    
-            <a href="{{url_for('edit_role', role_id = role.id)}}">Editar</a>
+            <a href="{{url_for('roles.edit_role', role_id = role.id)}}">Editar</a>
             {%endif%}
 
             {%if current_user.role.can_delete_roles and current_user.role != role %}
-            <a href="">Eliminar</a>
+            <a href="{{url_for('roles.delete_role', role_id = role.id)}}">Eliminar</a>
             {%endif%}
           </p>
         </div>
@@ -42,7 +42,7 @@
         <!-- New Post -->
         {%if current_user.role.can_create_roles%}
         <div class="d-flex justify-content-end mb-4">
-          <a class="btn btn-primary float-right" href="{{url_for('create_role')}}">Crear nuevo rol</a>
+          <a class="btn btn-primary float-right" href="{{url_for('roles.create_role')}}">Crear nuevo rol</a>
         </div>
         {%endif%}
       </div>
